chore: move settings to own util file
