+++ src/util/settings.ts
@@ -1,0 +1,3 @@
+import Gio from "gi://Gio";
+
+export const Settings = new Gio.Settings({ schema: pkg.name });
